+++ dwds/lib/src/debugging/debugger.dart
@@ -10,12 +10,13 @@ import 'package:meta/meta.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart'
     hide StackTrace;
 
+import '../loaders/strategy.dart';
 import '../readers/asset_reader.dart';
 import '../services/chrome_proxy_service.dart';
 import '../utilities/conversions.dart';
 import '../utilities/dart_uri.dart';
 import '../utilities/domain.dart';
-import '../utilities/objects.dart';
+import '../utilities/objects.dart' show Property;
 import '../utilities/shared.dart';
 import '../utilities/wrapped_service.dart';
 import 'dart_scope.dart';
@@ -81,6 +82,8 @@ class Debugger extends Domain {
   // TODO(alanknight): It would be nice to keep these as CallFrame instances,
   // but they don't map enough of the data yet.
   List<Map<String, dynamic>> _pausedJsStack;
+
+  List<Map<String, dynamic>> getJsStack() => _pausedJsStack;
 
   bool _isStepping = false;
 
@@ -386,7 +389,7 @@ class Debugger extends Domain {
     // want. To make those alternatives easier in JS, pass both count and end.
     var expression = '''
         function (offset, count, end) {
-          const sdk = $loadModule("dart_sdk");
+          const sdk = ${globalLoadStrategy.loadModuleSnippet}("dart_sdk");
           if (sdk.core.Map.is(this)) {
             const entries = sdk.dart.dload(this, "entries");
             const skipped = sdk.dart.dsend(entries, "skip", [offset])