+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -10,7 +10,6 @@ import 'package:pedantic/pedantic.dart';
 import 'package:pub_semver/pub_semver.dart' as semver;
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
-import '../../dwds.dart' show LogWriter;
 import '../connections/app_connection.dart';
 import '../debugging/debugger.dart';
 import '../debugging/execution_context.dart';
@@ -18,10 +17,14 @@ import '../debugging/inspector.dart';
 import '../debugging/location.dart';
 import '../debugging/modules.dart';
 import '../debugging/remote_debugger.dart';
+import '../loaders/strategy.dart';
 import '../readers/asset_reader.dart';
 import '../utilities/dart_uri.dart';
 import '../utilities/shared.dart';
 import '../utilities/wrapped_service.dart';
+
+import 'expression_compiler.dart';
+import 'expression_evaluator.dart';
 
 /// Adds [event] to the stream with [streamId] if there is anybody listening
 /// on that stream.
@@ -78,16 +81,19 @@ class ChromeProxyService implements VmServiceInterface {
   final _disabledBreakpoints = <Breakpoint>{};
   final _previousBreakpoints = <Breakpoint>{};
 
+  final LogWriter _logWriter;
+  ExpressionEvaluator _expressionEvaluator;
+
   ChromeProxyService._(
-    this._vm,
-    this.uri,
-    this._assetReader,
-    this.remoteDebugger,
-    this._modules,
-    this._locations,
-    this._restoreBreakpoints,
-    this.executionContext,
-  ) {
+      this._vm,
+      this.uri,
+      this._assetReader,
+      this.remoteDebugger,
+      this._modules,
+      this._locations,
+      this._restoreBreakpoints,
+      this.executionContext,
+      this._logWriter) {
     _debuggerCompleter.complete(Debugger.create(
       remoteDebugger,
       _streamNotify,
@@ -100,14 +106,14 @@ class ChromeProxyService implements VmServiceInterface {
   }
 
   static Future<ChromeProxyService> create(
-    RemoteDebugger remoteDebugger,
-    String tabUrl,
-    AssetReader assetReader,
-    AppConnection appConnection,
-    LogWriter logWriter,
-    bool restoreBreakpoints,
-    ExecutionContext executionContext,
-  ) async {
+      RemoteDebugger remoteDebugger,
+      String tabUrl,
+      AssetReader assetReader,
+      AppConnection appConnection,
+      LogWriter logWriter,
+      bool restoreBreakpoints,
+      ExecutionContext executionContext,
+      ExpressionCompiler expressionCompiler) async {
     // TODO: What about `architectureBits`, `targetCPU`, `hostCPU` and `pid`?
     final vm = VM()
       ..isolates = []
@@ -117,9 +123,23 @@ class ChromeProxyService implements VmServiceInterface {
     var modules = Modules(remoteDebugger, tabUrl, executionContext);
     var locations = Locations(assetReader, modules, tabUrl);
     var service = ChromeProxyService._(vm, tabUrl, assetReader, remoteDebugger,
-        modules, locations, restoreBreakpoints, executionContext);
+        modules, locations, restoreBreakpoints, executionContext, logWriter);
+
     unawaited(service.createIsolate(appConnection));
+    await service.createEvaluator(expressionCompiler);
     return service;
+  }
+
+  /// Creates expression evaluator to use in [evaluateInFrame]
+  ///
+  /// Expression evaluation is only supported with scenarios that
+  /// provide non-null [ExpressionCompiler] to [create].
+  /// Otherwise [evaluateInFrame] will throw unsupported exception.
+  Future<void> createEvaluator(ExpressionCompiler compiler) async {
+    _expressionEvaluator = compiler == null
+        ? null
+        : ExpressionEvaluator(
+            await _debugger, _locations, _modules, compiler, _logWriter);
   }
 
   /// Creates a new isolate.
@@ -262,7 +282,7 @@ class ChromeProxyService implements VmServiceInterface {
     var stringArgs = args.map((k, v) => MapEntry(
         k is String ? k : jsonEncode(k), v is String ? v : jsonEncode(v)));
     var expression = '''
-$loadModule("dart_sdk").developer.invokeExtension(
+${globalLoadStrategy.loadModuleSnippet}("dart_sdk").developer.invokeExtension(
     "$method", JSON.stringify(${jsonEncode(stringArgs)}));
 ''';
     var response =
@@ -302,8 +322,22 @@ $loadModule("dart_sdk").developer.invokeExtension(
 
   @override
   Future evaluateInFrame(String isolateId, int frameIndex, String expression,
-      {Map<String, String> scope, bool disableBreakpoints}) {
-    throw UnimplementedError();
+      {Map<String, String> scope, bool disableBreakpoints}) async {
+    if (_expressionEvaluator != null) {
+      var isolate = _inspector?.isolate;
+      if (isolate?.id != isolateId) {
+        throw ArgumentError.value(
+            isolateId, 'isolateId', 'Unrecognized isolate id');
+      }
+
+      var result = await _expressionEvaluator.evaluateExpression(
+          isolateId, frameIndex, expression);
+
+      return _inspector?.instanceHelper?.instanceRefFor(result);
+    }
+
+    throw UnimplementedError(
+        'Expression evaluation is not supported for this configuration');
   }
 
   @override