+++ dwds/web/reloader/require_restarter.dart
@@ -14,9 +14,7 @@ import 'package:graphs/graphs.dart' as graphs;
 import 'package:js/js.dart';
 import 'package:js/js_util.dart';
 import 'package:path/path.dart' as p;
-import 'package:pedantic/pedantic.dart';
 
-import '../module.dart';
 import '../promise.dart';
 import 'restarter.dart';
 
@@ -31,15 +29,15 @@ external DartLoader get dartLoader;
 @JS(r'$loadModuleConfig')
 external Object Function(String module) get require;
 
+@JS(r'$dartRunMain')
+external set dartRunMain(Function() func);
+
 List<K> keys<K, V>(JsMap<K, V> map) {
   return List.from(_jsArrayFrom(map.keys()));
 }
 
 @JS('Array.from')
 external List _jsArrayFrom(Object any);
-
-@JS('Object.keys')
-external List _jsObjectKeys(Object any);
 
 @JS('Object.values')
 external List _jsObjectValues(Object any);
@@ -59,54 +57,6 @@ class DartLoader {
   @JS()
   external void forceLoadModule(String moduleId, void Function() callback,
       void Function(JsError e) onError);
-
-  @JS()
-  external Object getModuleLibraries(String moduleId);
-}
-
-@anonymous
-@JS()
-abstract class HotReloadableLibrary {
-  /// Implement this function to handle update of child modules.
-  ///
-  /// May return nullable bool. To indicate that reload of child completes
-  /// successfully return true. To indicate that hot-reload is undoable for this
-  /// child return false - this will lead to full page reload. If null returned,
-  /// reloading will be propagated to current module itself.
-  ///
-  /// The name of the child will be provided in [childId]. New version of child
-  /// module object will be provided in [child].
-  /// If any state was saved from previous version, it will be passed to [data].
-  ///
-  /// This function will be called on old version of module current after child
-  /// reloading.
-  @JS()
-  external bool hot$onChildUpdate(String childId, HotReloadableLibrary child,
-      [Object data]);
-
-  /// Implement this function with any code to release resources before destroy.
-  ///
-  /// Any object returned from this function will be passed to update hooks. Use
-  /// it to save any state you need to be preserved between hot reloadings.
-  /// Try do not use any custom types here, as it might prevent their code from
-  /// reloading. Better serialise to JSON or plain types.
-  ///
-  /// This function will be called on old version of module before unloading.
-  @JS()
-  external Object hot$onDestroy();
-
-  /// Implement this function to handle update of the module itself.
-  ///
-  /// May return nullable bool. To indicate that reload completes successfully
-  /// return true. To indicate that hot-reload is undoable return false - this
-  /// will lead to full page reload. If null returned, reloading will be
-  /// propagated to parent.
-  ///
-  /// If any state was saved from previous version, it will be passed to [data].
-  ///
-  /// This function will be called on new version of module after reloading.
-  @JS()
-  external bool hot$onSelfUpdate([Object data]);
 }
 
 class HotReloadFailedException implements Exception {
@@ -133,39 +83,6 @@ abstract class JsMap<K, V> {
 
   @JS()
   external Object keys();
-}
-
-class LibraryWrapper implements Library {
-  final HotReloadableLibrary _internal;
-
-  LibraryWrapper(this._internal);
-
-  @override
-  bool onChildUpdate(String childId, Library child, [Object data]) {
-    if (_internal != null && hasProperty(_internal, r'hot$onChildUpdate')) {
-      return _internal.hot$onChildUpdate(
-          childId, (child as LibraryWrapper)._internal, data);
-    }
-    // ignore: avoid_returning_null
-    return null;
-  }
-
-  @override
-  Object onDestroy() {
-    if (_internal != null && hasProperty(_internal, r'hot$onDestroy')) {
-      return _internal.hot$onDestroy();
-    }
-    return null;
-  }
-
-  @override
-  bool onSelfUpdate([Object data]) {
-    if (_internal != null && hasProperty(_internal, r'hot$onSelfUpdate')) {
-      return _internal.hot$onSelfUpdate(data);
-    }
-    // ignore: avoid_returning_null
-    return null;
-  }
 }
 
 /// Handles hot restart reloading for use with the require module system.
@@ -216,10 +133,7 @@ class RequireRestarter implements Restarter {
     if (modulesToLoad.isNotEmpty) {
       _updateGraph();
       var result = await _reload(modulesToLoad);
-      if (result == null) {
-        callMethod(getProperty(require('dart_sdk'), 'dart'), 'hotRestart', []);
-        result = true;
-      }
+      callMethod(getProperty(require('dart_sdk'), 'dart'), 'hotRestart', []);
       return result;
     }
 
@@ -237,20 +151,6 @@ class RequireRestarter implements Restarter {
 
   Future<void> _initialize() async {
     _lastKnownDigests = await _getDigests();
-  }
-
-  Module _moduleLibraries(String moduleId) {
-    var moduleObj = dartLoader.getModuleLibraries(moduleId);
-    if (moduleObj == null) {
-      throw HotReloadFailedException("Failed to get module '$moduleId'. "
-          "This error might appear if such module doesn't exist or isn't already "
-          'loaded');
-    }
-    var moduleKeys = List<String>.from(_jsObjectKeys(moduleObj));
-    var moduleValues =
-        List<HotReloadableLibrary>.from(_jsObjectValues(moduleObj));
-    var moduleLibraries = moduleValues.map((x) => LibraryWrapper(x));
-    return Module(Map.fromIterables(moduleKeys, moduleLibraries));
   }
 
   List<String> _moduleParents(String module) =>
@@ -272,8 +172,8 @@ class RequireRestarter implements Restarter {
         Comparable.compare(_moduleOrdering[module2], _moduleOrdering[module1]);
 
     if (topological == 0) {
-      // If modules are in cycle (same strongly connected component) compare their
-      // string id, to ensure total ordering for SplayTreeSet uniqueness.
+      // If modules are in cycle (same strongly connected component) compare
+      // their string id, to ensure total ordering for SplayTreeSet uniqueness.
       topological = module1.compareTo(module2);
     }
 
@@ -291,78 +191,45 @@ class RequireRestarter implements Restarter {
     if (!_running.isCompleted) return await _running.future;
     _running = Completer();
 
-    // We want to schedule some async work for the future but return the
-    // `_running` completers future synchronously.
-    unawaited(() async {
-      var reloadedModules = 0;
-
-      try {
-        while (_dirtyModules.isNotEmpty) {
-          var moduleId = _dirtyModules.first;
-          _dirtyModules.remove(moduleId);
+    var reloadedModules = 0;
+    try {
+      String previousModuleId;
+      while (_dirtyModules.isNotEmpty) {
+        var moduleId = _dirtyModules.first;
+        _dirtyModules.remove(moduleId);
+        var parentIds = _moduleParents(moduleId);
+        // Check if this is the root / bootstrap module.
+        if (parentIds == null || parentIds.isEmpty) {
+          // The bootstrap module is not reloaded but we need to update the
+          // $dartRunMain reference to the newly loaded child module.
+          var childModule = callMethod(getProperty(require('dart_sdk'), 'dart'),
+              'getModuleLibraries', [previousModuleId]);
+          dartRunMain = allowInterop(() {
+            callMethod(_jsObjectValues(childModule).first, 'main', []);
+          });
+        } else {
           ++reloadedModules;
-
-          var existing = _moduleLibraries(moduleId);
-          var data = existing.onDestroy();
-
-          var newVersion = await _reloadModule(moduleId);
-          var success = newVersion.onSelfUpdate(data);
-          if (success == true) continue;
-          if (success == false) {
-            print("Module '$moduleId' is marked as unreloadable. "
-                'Firing full page reload.');
-            _reloadPage();
-            return false;
-          }
-
-          var parentIds = _moduleParents(moduleId);
-          if (parentIds == null || parentIds.isEmpty) {
-            print("Module reloading wasn't handled by any of parents. "
-                'Firing full page reload.');
-            _reloadPage();
-            return false;
-          }
+          await _reloadModule(moduleId);
           parentIds.sort(_moduleTopologicalCompare);
-
-          for (var parentId in parentIds) {
-            var parentModule = _moduleLibraries(parentId);
-            success = parentModule.onChildUpdate(moduleId, newVersion, data);
-
-            // Found the bootstrap module at the top of the app, return null to
-            // indicate an unhandled reload.
-            if (parentId.endsWith('.dart.bootstrap')) {
-              return null;
-            }
-
-            if (success == true) continue;
-            if (success == false) {
-              print("Module '$moduleId' is marked as unreloadable. "
-                  'Firing full page reload.');
-              _reloadPage();
-              return false;
-            }
-            _dirtyModules.add(parentId);
-          }
+          _dirtyModules.addAll(parentIds);
+          previousModuleId = moduleId;
         }
-        print('$reloadedModules modules were hot-reloaded.');
-      } on HotReloadFailedException catch (e) {
-        print('Error during script reloading. Firing full page reload. $e');
-        _reloadPage();
-        return false;
       }
-      return true;
-    }()
-        .then(_running.complete)
-        .catchError(_running.completeError));
-
+      print('$reloadedModules module(s) were hot-reloaded.');
+      _running.complete(true);
+    } on HotReloadFailedException catch (e) {
+      print('Error during script reloading. Firing full page reload. $e');
+      _reloadPage();
+      _running.complete(false);
+    }
     return _running.future;
   }
 
-  Future<Module> _reloadModule(String moduleId) {
-    var completer = Completer<Module>();
+  Future<void> _reloadModule(String moduleId) {
+    var completer = Completer();
     var stackTrace = StackTrace.current;
     dartLoader.forceLoadModule(moduleId, allowInterop(() {
-      completer.complete(_moduleLibraries(moduleId));
+      completer.complete();
     }),
         allowInterop((e) => completer.completeError(
             HotReloadFailedException(e.message), stackTrace)));